backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.0.0
numpy==1.26.4
opencv-python==4.8.1.78